requirements.txt
FinanceDataReader
requests
pandas
numpy
python-dotenv
